How to access the dynamically loaded svg
As discussed with @florianletsch on gitter. Note that for static icons, colour change can be done more easily by adding a css rule in the main stylesheet, but this can't be done if the icon is dynamic and the colour is altered programmatically. There may be other uses for adjusting the svg DOM too.