Add meta viewport in Basic Html Markup
It´s not UIkit subject, but since it´s essential for the responsive design to work, add media viewport to basic html structure.
Bulma have it, Bootstrap have it, Foundation too. 
This prevents a bad experience for new user making first steps in responsive design, i see it quite ofter in Gitter.