Update javascript.md
Correct me if I'm wrong but after I updated from 3.1.9 to 3.2.1 this is clearly not supported anymore. I tried initializing modals, drops and filters. I observe that only a selector or a DOM node is accepted to initialize the component. Additionally I also observed the same behavior with passing a component to UIkit.util.on(). It can be only a selector or a node. It's a pity though because it would be really handy that one could as well pass jQuery object or even a component ref itself.